Move helper fns from omexml to utils
