+++ src/loaders/utils.ts
@@ -94,3 +94,25 @@ export function getChannelStats(arr: TypedArray) {
     autoSliders
   };
 }
+
+export function ensureArray<T>(x: T | T[]) {
+  return Array.isArray(x) ? x : [x];
+}
+
+type RGBA = [r: number, g: number, b: number, a: number];
+
+// Adapted from: https://github.com/ome/ome-zarr-py/blob/db60b8272e0fe005920f8a296d4828b7a32e663e/ome_zarr/conversions.py#L16
+export function intToRgba(int: number): RGBA {
+  if (!Number.isInteger(int)) {
+    throw Error('Not an integer.');
+  }
+
+  // Write number to int32 representation (4 bytes).
+  const buffer = new ArrayBuffer(4);
+  const view = new DataView(buffer);
+  view.setInt32(0, int, false); // offset === 0, littleEndian === false
+
+  // Take u8 view and extract number for each byte (1 byte for R/G/B/A).
+  const bytes = new Uint8Array(buffer);
+  return Array.from(bytes) as RGBA;
+}
